Move authz construction to reloader
